Add BSP Tree

+++ wasm/include/GLTF.h
@@ -171,7 +171,7 @@ class GLTF : public OptimizationProblem, public TableInterface {
         void setTetraModel(glm::vec3 center, float size);
 
         // Sets the model to a polyhedron of the given color (Can be used to generate visuals for ConvexShape objects)
-        void setPolyhedronModel(std::vector<glm::vec3>& vertices, std::vector<std::vector<int>>& faces, glm::vec3 color);
+        void setPolyhedronModel(std::vector<glm::dvec3>& vertices, std::vector<std::vector<int>>& faces, glm::vec3 color);
 
         void addPrimitive(std::vector<Vertex>& vertices, std::vector<Triangle>& triangles,
             Variant& primitive, int node_id, const glm::mat4& transform, Variant& json, const Variant& bin);
